I think this is gonna work man!
Check out the training scene. It looks a bit fucked at the moment because it doesn't like having multiple frogs (you'll notice that the animations are stuffed) but whatevs, can fix that later.

I wrote a really quick and dirty feed forward neural net and a steering behaviour that uses the net to derive directions. I had to tweak quite a few parameters to get the frog even vaguely moving sensibly. There's no learning at all at the moment, but you'll notice that even so, one of the eight initial frogs will have a tendency to target the flies. (It's *essentially* just feeding in the position of the nearest fly at the moment.) It should be pretty easy to start learning the neural net based on the GA stuff you've written. Nice work on that, looks profesh!